Version 1.0 release!
Finalized all the macros, updated the readme and added a link to it in the changelog. All macros linked, all journal entries linked. The entire module can now be used from the compendia directly, but dragging the WR Weather Reports Generator macro to the hotbar is recommended. Simple Calendar integration is complete. FXMaster integration is complete.