v1.0.1 Bug Fixes
Bug fixes:
Jungle Rolltables - Had extra results left over from an earlier version. The results have been removed and the jungle roll tables are fixed.
Forest Rolltables - Compendium UUID for Forest Spring was pointing to Jungle Spring. Correct UUID has ben entered and now rolls Forest Spring when requested.
Readme - Didn't realize the images pasted into the journal were being saved as a png on the assests folder. Moved the images to the module /img/readme and correctly linked them in the readme.
Readme - Added that Simple Calendar is required, not just recommended, for this current version and my be removed as a requirement in a future version.
Added Simple Calendar as a dependency.